pyproject.toml
[project]
name = "dismissal-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.119.0",
    "pwdlib[argon2]>=0.2.1",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.6",
    "sqlalchemy>=2.0.44",
    "taskipy>=1.14.1",
]

[tool.ruff]
line-length = 79

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT', 'FAST']

[tool.ruff.format]
preview = true

[tool.pytest.ini_options]
pythonpath = "."

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'fastapi dev app/main.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=app -vv'
post_test = 'coverage html'